forms_api/requirements.txt
django
djangorestframework
Pillow
XlsxWriter
django-rest-swagger
pyyaml
drf-yasg
django-rest-registration 